adding validation code separate from CLI to make cli more readable

+++ src/main/java/scraping/models/SectionAttribute.java
@@ -16,8 +16,8 @@ public class SectionAttribute {
   private String description;
   private String instructionMode;
   private String instructor; // dup
-  private float minUnits;
-  private float maxUnits;
+  private Double minUnits;
+  private Double maxUnits;
   private String grading;
   private String room;
   private String prerequisites;
@@ -26,7 +26,7 @@ public class SectionAttribute {
   public SectionAttribute(String courseName, int registrationNumber,
                           SectionStatus status, String campus,
                           String description, String instructorMode,
-                          String instructor, float minUnits, float maxUnits,
+                          String instructor, Double minUnits, Double maxUnits,
                           String grading, String prerequisites, String room) {
     this.courseName = courseName;
     this.registrationNumber = registrationNumber;
@@ -56,9 +56,9 @@ public class SectionAttribute {
 
   public @NotNull String getInstructor() { return instructor; }
 
-  public float getMinUnits() { return minUnits; }
+  public Double getMinUnits() { return minUnits; }
 
-  public float getMaxUnits() { return maxUnits; }
+  public Double getMaxUnits() { return maxUnits; }
 
   public String getGrading() { return grading; }
